my profil page

+++ src/Controller/IngredientController.php
@@ -8,7 +8,9 @@ use App\Repository\IngredientRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\PropertyAccess\Exception\AccessException;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 
 #[Route('/ingredient')]
 class IngredientController extends AbstractController
@@ -21,7 +23,7 @@ class IngredientController extends AbstractController
         ]);
     }
 
-    #[Route('/new', name: 'ingredient_new', methods: ['GET','POST'])]
+    #[Route('/new', name: 'ingredient_new', methods: ['GET', 'POST'])]
     public function new(Request $request): Response
     {
         $ingredient = new Ingredient();
@@ -29,6 +31,7 @@ class IngredientController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            $ingredient->setAuthor($this->getUser());
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->persist($ingredient);
             $entityManager->flush();
@@ -50,9 +53,14 @@ class IngredientController extends AbstractController
         ]);
     }
 
-    #[Route('/{id}/edit', name: 'ingredient_edit', methods: ['GET','POST'])]
+    #[Route('/{id}/edit', name: 'ingredient_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Ingredient $ingredient): Response
     {
+
+        if ($ingredient->getAuthor() !== $this->getUser()) {
+            throw new AccessDeniedException();
+        }
+
         $form = $this->createForm(IngredientType::class, $ingredient);
         $form->handleRequest($request);
 
@@ -71,7 +79,12 @@ class IngredientController extends AbstractController
     #[Route('/{id}', name: 'ingredient_delete', methods: ['POST'])]
     public function delete(Request $request, Ingredient $ingredient): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$ingredient->getId(), $request->request->get('_token'))) {
+
+        if ($ingredient->getAuthor() !== $this->getUser()) {
+            throw new AccessDeniedException();
+        }
+
+        if ($this->isCsrfTokenValid('delete' . $ingredient->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($ingredient);
             $entityManager->flush();
